enseada-server is now a prettier CLI tool with usage and (some) flags
