+++ cmd/enseada-server/main.go
@@ -14,159 +14,133 @@ import (
 	"syscall"
 	"time"
 
-	"cloud.google.com/go/errorreporting"
-	"github.com/airbrake/gobrake/v4"
-	enseada "github.com/enseadaio/enseada/pkg"
-	"github.com/enseadaio/enseada/pkg/errare"
-	"github.com/enseadaio/enseada/pkg/errare/airbrake"
-	logerr "github.com/enseadaio/enseada/pkg/errare/log"
-	sentryerr "github.com/enseadaio/enseada/pkg/errare/sentry"
-	"github.com/enseadaio/enseada/pkg/errare/stackdriver"
-	"github.com/getsentry/sentry-go"
-
 	"github.com/enseadaio/enseada/pkg/app"
-
 	"github.com/enseadaio/enseada/pkg/log/adapters"
+	"github.com/joho/godotenv"
+	"github.com/spf13/cobra"
+
 	"github.com/spf13/viper"
 
 	"github.com/enseadaio/enseada/pkg/log"
-	"github.com/joho/godotenv"
 )
 
-func init() {
-	if info, err := os.Stat("./.env"); err == nil && !info.IsDir() {
-		err := godotenv.Load()
+const (
+	defaultPort = 9623
+)
+
+var c = viper.NewWithOptions(
+	viper.EnvKeyReplacer(strings.NewReplacer(".", "_")),
+)
+
+var rootCmd = &cobra.Command{
+	Use:   "enseada-server",
+	Short: "A Cloud native multi-package registry",
+	Long: `A Cloud native multi-package registry
+	
+Enseada is a modern, fast and scalable package registry, designed from the ground up to run in elastic, container-based environments and to be highly available and distributed.
+More information available at https://enseada.io`,
+	PreRunE: func(cmd *cobra.Command, args []string) error {
+		if info, err := os.Stat("./.env"); err == nil && !info.IsDir() {
+			err := godotenv.Load()
+			if err != nil {
+				return err
+			}
+		}
+
+		c.SetDefault("log.level", "info")
+		c.SetDefault("port", defaultPort)
+		c.SetDefault("storage.provider", "local")
+		c.SetDefault("local.storage.dir", "uploads")
+		c.SetDefault("root.password", "root")
+		c.SetDefault("sentry.environment", "development")
+		c.SetDefault("airbrake.environment", "development")
+
+		c.RegisterAlias("aws.region", "s3.region")
+		c.RegisterAlias("aws.bucket", "s3.bucket")
+		c.RegisterAlias("aws.bucket.prefix", "s3.bucket.prefix")
+		c.RegisterAlias("aws.s3.endpoint", "s3.endpoint")
+		c.RegisterAlias("aws.s3.sse", "s3.sse")
+
+		c.AutomaticEnv()
+
+		return nil
+	},
+	Run: func(cmd *cobra.Command, args []string) {
+		lvl := log.Level(strings.ToLower(c.GetString("log.level")))
+		l, err := adapters.NewZapLoggerAdapter(lvl)
 		if err != nil {
 			panic(err)
 		}
-	}
+
+		l.Info("Enseada booting up...")
+		start := time.Now()
+
+		errh, err := errorHandler(l, c)
+		if err != nil {
+			l.Fatal(err)
+		}
+		defer func() {
+			if err := errh.Close(); err != nil {
+				l.Error(err)
+			}
+		}()
+
+		bootctx, cancelboot := context.WithTimeout(context.Background(), 10*time.Second)
+		defer cancelboot()
+
+		mods, err := modules(bootctx, l, c, errh)
+		if err != nil {
+			l.Fatal(err)
+		}
+
+		a := app.New(
+			app.Modules(mods...),
+			app.OnError(func(err error) {
+				errh.HandleError(err)
+			}),
+			app.OnPanic(func(v interface{}) {
+				errh.HandlePanic(v)
+			}),
+		)
+
+		ctx, cancel := context.WithCancel(context.Background())
+		if err := a.Start(ctx); err != nil {
+			l.Fatal(err)
+		}
+		l.Infof("started Enseada in %dms", time.Since(start).Milliseconds())
+
+		quit := make(chan os.Signal)
+		signal.Notify(quit, syscall.SIGINT, syscall.SIGKILL, syscall.SIGTERM, syscall.SIGQUIT)
+		<-quit
+
+		l.Info("terminating Enseada...")
+		defer l.Infof("stopped Enseada")
+
+		cancel()
+
+		shutdownctx, cancelshutdown := context.WithTimeout(context.Background(), 5*time.Second)
+		defer cancelshutdown()
+
+		if err := a.Stop(shutdownctx); err != nil {
+			l.Fatal(err)
+		}
+	},
 }
 
-func conf() *viper.Viper {
-	c := viper.NewWithOptions(
-		viper.EnvKeyReplacer(strings.NewReplacer(".", "_")),
-	)
+func init() {
+	// Subcommands
+	rootCmd.AddCommand(versionCmd)
 
-	c.SetDefault("log.level", "info")
-	c.SetDefault("port", "9623")
-	c.SetDefault("storage.provider", "local")
-	c.SetDefault("local.storage.dir", "uploads")
-	c.SetDefault("root.password", "root")
-	c.SetDefault("sentry.environment", "development")
-	c.SetDefault("airbrake.environment", "development")
+	// Flags
+	rootCmd.Flags().IntP("port", "p", defaultPort, "HTTP port to use")
+	rootCmd.Flags().StringP("log", "l", string(log.INFO), "logging level )")
 
-	c.RegisterAlias("aws.region", "s3.region")
-	c.RegisterAlias("aws.bucket", "s3.bucket")
-	c.RegisterAlias("aws.bucket.prefix", "s3.bucket.prefix")
-	c.RegisterAlias("aws.s3.endpoint", "s3.endpoint")
-	c.RegisterAlias("aws.s3.sse", "s3.sse")
-
-	c.AutomaticEnv()
-	return c
-}
-
-func errorHandler(logger log.Logger, c *viper.Viper) (errare.Handler, error) {
-	l := logerr.NewHandler(logger, true)
-
-	var erh errare.Handler
-	reporter := strings.ToLower(c.GetString("error.reporter"))
-	switch reporter {
-	case "sentry":
-		sen, err := sentryerr.NewHandler(sentry.ClientOptions{
-			Dsn:              c.GetString("sentry.dsn"),
-			Debug:            false,
-			AttachStacktrace: true,
-			Environment:      c.GetString("sentry.environment"),
-		})
-		if err != nil {
-			return nil, err
-		}
-		erh = sen
-	case "stackdriver":
-		sd, err := stackdriver.NewHandler(context.Background(), c.GetString("google.project.id"), errorreporting.Config{
-			ServiceName:    "enseada",
-			ServiceVersion: enseada.VersionString(),
-			OnError: func(err error) {
-				l.HandleError(err)
-			},
-		})
-		if err != nil {
-			return nil, err
-		}
-		erh = sd
-	case "airbrake":
-		erh = airbrake.NewHandler(&gobrake.NotifierOptions{
-			ProjectId:   c.GetInt64("airbrake.project.id"),
-			ProjectKey:  c.GetString("airbrake.project.key"),
-			Environment: c.GetString("airbrake.environment"),
-		})
-	default:
-		logger.Info("no error reporter configured. Defaulting to log")
-		return l, nil
-	}
-
-	logger.Infof("configured error reporting with %s", reporter)
-	return errare.Compose(l, erh), nil
+	c.BindPFlag("port", rootCmd.Flags().Lookup("port"))
+	c.BindPFlag("log.level", rootCmd.Flags().Lookup("log"))
 }
 
 func main() {
-	c := conf()
-	lvl := log.Level(strings.ToLower(c.GetString("log.level")))
-	l, err := adapters.NewZapLoggerAdapter(lvl)
-	if err != nil {
+	if err := rootCmd.Execute(); err != nil {
 		panic(err)
-	}
-
-	l.Info("Enseada booting up...")
-	start := time.Now()
-
-	errh, err := errorHandler(l, c)
-	if err != nil {
-		l.Fatal(err)
-	}
-	defer func() {
-		if err := errh.Close(); err != nil {
-			l.Error(err)
-		}
-	}()
-
-	bootctx, cancelboot := context.WithTimeout(context.Background(), 10*time.Second)
-	defer cancelboot()
-
-	mods, err := modules(bootctx, l, c, errh)
-	if err != nil {
-		l.Fatal(err)
-	}
-
-	a := app.New(
-		app.Modules(mods...),
-		app.OnError(func(err error) {
-			errh.HandleError(err)
-		}),
-		app.OnPanic(func(v interface{}) {
-			errh.HandlePanic(v)
-		}),
-	)
-
-	ctx, cancel := context.WithCancel(context.Background())
-	if err := a.Start(ctx); err != nil {
-		l.Fatal(err)
-	}
-	l.Infof("started Enseada in %dms", time.Since(start).Milliseconds())
-
-	quit := make(chan os.Signal)
-	signal.Notify(quit, syscall.SIGINT, syscall.SIGKILL, syscall.SIGTERM, syscall.SIGQUIT)
-	<-quit
-
-	l.Info("terminating Enseada...")
-	defer l.Infof("stopped Enseada")
-
-	cancel()
-
-	shutdownctx, cancelshutdown := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancelshutdown()
-
-	if err := a.Stop(shutdownctx); err != nil {
-		l.Fatal(err)
 	}
 }
