+++ cmd/enseada-server/version.go
